Add files via upload
This report discusses the complexities and challenges associated with intellectual property (IP) licensing for NFTs that are used in Picture for Profile (PFP) Projects. Since 2017, PFP Projects have used a variety of IP licenses to exert control over the art associated with the NFT, while still empowering the NFT holders. The premise of this research is that the lack of an industry-standard slows innovation and harms consumers.